chore(db): move non-prod prisma stuff to separate directory
