+++ libs/db/src/scripts/reset.ts
@@ -1,4 +1,4 @@
-import { nuke } from '../prisma/utils';
+import { nuke } from '../prisma-utils/utils';
 import { prismaWrapper } from './prisma-wrapper';
 
 prismaWrapper(async (prisma) => await nuke(prisma));
